Bug 1437570 - Upgrade to rust-ini 0.10.2. r=jgraham
Upgrading to rust-ini 0.10.2 has the benefit that it no longers
depends on a too specific version of the log crate.  We currently
compile two different versions of log as part of the geckodriver
build, and it will marginally increase compile performance not to
compile that twice.

MozReview-Commit-ID: HAwa4Kg8Lyn

--HG--
extra : rebase_source : d81b3450934f011663b508271c8c6a0f92997490