Bug 1494637 - [geckodriver] Remove custom serialization of JavascriptCommandParameters for Marionette. r=ato
The argument "specialPowers" is no longer used in Marionette
and as such can be removed from the serialization.

The default value for "newSandbox" in Marionette is false,
so it doesn't need to be explicitely set in geckodriver.

By removing those two properties a custom serialization
is no longer neccessary.

Depends on D7210

Differential Revision: https://phabricator.services.mozilla.com/D7211

--HG--
extra : moz-landing-system : lando