Bug 1509256 - Clarify where to add preferences for geckodriver and Marionette. r=ato
It should be made clear that only those preferences should be added
to marionette.js which have an immediate effect. All others which
eg. require a restart, or have to be set before the application
starts the first time, must be added to the appropriate client's
profile generation code.

Differential Revision: https://phabricator.services.mozilla.com/D13059

--HG--
extra : moz-landing-system : lando