bug 1526938: geckodriver: generalise cli errors; r=whimboo
Instead of returning a tuple of (i32, String), causing us to lose
details and the cause of an error, this patch formalises the error
scenarios that may occur during command-line parsing and startup
of the WebDriver server.

Exit codes are encoded with each variant of FatalError, instead of
being specified in each case.

Type conversions are implemented for clap::Error (flag parsing)
and io::Error (webdriver::server).

Depends on D19426

Differential Revision: https://phabricator.services.mozilla.com/D19427

--HG--
extra : moz-landing-system : lando