bug 1526938: geckodriver: fix missing flag parsing errors; r=whimboo
Because we emit the flag parsing errors to the log, through error!(),
they are subject to whether the logging subsystem is enabled.
Because logging is disabled by default, no error information is
currently displayed to the user.

Since we cannot initialise logging implicitly due to the risk of
emitting log messages the user did not request, this patch changes
geckodriver to print the flag parsing errors to stderr

Depends on D19429

Differential Revision: https://phabricator.services.mozilla.com/D19430

--HG--
extra : moz-landing-system : lando