bug 1526938: geckodriver: display help message on non-clap errors; r=whimboo
The help message is implicitly included in clap error descriptions, but
they are not for errors originating from this file.  By introducing
a FatalError::help_included() function we make sure we print the
help message in all cases.

An alternative implementation, which perhaps would be more idiomatic,
would be to prepare the help message within the fmt::Display trait
implementation for FatalError, but I could find no way of passing
in a reference to clap::App without storing it in an atomic global const.

Depends on D19430

Differential Revision: https://phabricator.services.mozilla.com/D19431

--HG--
extra : moz-landing-system : lando