Bug 1577850 - Use GeckoView configuration YAML file in geckodriver. r=nalexander,webdriver-reviewers,jgraham
This makes geckodriver configure GeckoView vehicles using a
configuration YAML file.  This is the [blessed way to configure such
vehicles](https://mozilla.github.io/geckoview/consumer/docs/automation).

The old way based on intent arguments is kept as long as Fennec
needs to be supported because it doesn't read the configuration file.
And other packages, which are based on GeckoView, will ignore
package specific intent arguments.

The configuration file is always pushed, even when it will be
ignored by the relevant package.

Differential Revision: https://phabricator.services.mozilla.com/D44578

--HG--
extra : moz-landing-system : lando