Bug 1686707 - Support writing prefs to a command-line provided profile, r=webdriver-reviewers,whimboo
Previously we would just ignore the command-line provided profile,
write prefs to a new profile and then use the command line provided
one. That doesn't really make sense.

Now instead we try to write the prefs to the provided profile and
restore it after the test run.



Depends on D101771

Differential Revision: https://phabricator.services.mozilla.com/D101772